requirements.txt
streamlit==1.28.1
pandas==2.1.3
numpy==1.24.3
scikit-learn==1.3.2
xgboost==2.0.2
matplotlib==3.8.2
seaborn==0.13.0
opencv-python==4.8.1.78
pytesseract==0.3.10
Pillow==10.1.0
reportlab==4.0.7
plotly==5.17.0
requests==2.31.0
regex==2023.10.3
nltk==3.8.1
spacy==3.7.2
